--- manifest/v1alpha/slo/composite_test.go
@@ -1,1 +1,0 @@
-package slo
