+++ manifest/v1alpha/slo/test_data/expected_string_is_dns_subdomain_error.txt
@@ -1,0 +1,1 @@
+string does not match regular expression: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?$' (e.g. 'my-name', '123-abc'); a DNS-1123 compliant name must consist of lower case alphanumeric characters or '-', and must start and end with an alphanumeric character
